+++ CHANGELOG.md
@@ -6,8 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - `ViveGrip_GripPoint` is now solely responsible for messaging instead of acting on state change
 - Moved the tint color to `ViveGrip_Highlighter`
+- TODO: Modified `ViveGripExample_Hand` to use highlight triggers instead of touch
 
 ### Added
+- `ViveGripHighlightStart` and `ViveGripHighlightStop`
 - `ViveGrip_Grabber` now does the grabbing functionality that the grip point used to do
 
 ## [2.3.0] - 2016-07-16